+++ includes/delete_tagline_image.php
@@ -15,9 +15,9 @@ if (isset($_POST['article_id']) && $_POST['article_id'] != 0)
 {
 	$db->sqlquery("SELECT `article_top_image`,`article_top_image_filename`,`title`,`tagline_image` FROM `articles` WHERE `article_id` = ?", array($_POST['article_id']));
 	$article = $db->fetch();
-			
+
 	// remove old image
-	if ($article['article_top_image'] == 1)
+	if ($article['article_top_image'] == 1 && !empty($article['article_top_image_filename']))
 	{
 		if (unlink($_SERVER['DOCUMENT_ROOT'] . '/uploads/articles/topimages/' . $article['article_top_image_filename']))
 		{
@@ -31,7 +31,7 @@ if (isset($_POST['article_id']) && $_POST['article_id'] != 0)
 			$check = 1;
 		}
 	}
-						
+
 	$db->sqlquery("UPDATE `articles` SET `article_top_image` = 0, `article_top_image_filename` = '', `tagline_image` = '' WHERE `article_id` = ?", array($_POST['article_id']));
 
 	if ($check == 1)
@@ -46,15 +46,18 @@ if (isset($_POST['article_id']) && $_POST['article_id'] != 0)
 }
 else
 {
-	if(unlink($_SERVER['DOCUMENT_ROOT'] . '/uploads/articles/tagline_images/' . $_SESSION['uploads_tagline']['image_name']) && unlink($_SERVER['DOCUMENT_ROOT'] . '/uploads/articles/tagline_images/thumbnails/' . $_SESSION['uploads_tagline']['image_name']))
+	if (isset($_SESSION['uploads_tagline']['image_name']))
 	{
-		unset($_SESSION['uploads_tagline']);
-		echo "YES";
-	}
+		if(unlink($_SERVER['DOCUMENT_ROOT'] . '/uploads/articles/tagline_images/' . $_SESSION['uploads_tagline']['image_name']) && unlink($_SERVER['DOCUMENT_ROOT'] . '/uploads/articles/tagline_images/thumbnails/' . $_SESSION['uploads_tagline']['image_name']))
+		{
+			unset($_SESSION['uploads_tagline']);
+			echo "YES";
+		}
 
-	else
-	{
-		echo "NO";
+		else
+		{
+			echo "NO";
+		}
 	}
 }
 ?>
